+++ src/main/MouseInput.java
@@ -89,11 +89,11 @@ public class MouseInput{
 		}
 		
 		public int getMapX() {
-			return (int)(x/Screen.MAP_SCALE/Screen.MAP_ZOOM+Screen.xOffset);
+			return (int)(((double)x)/Screen.MAP_SCALE/Screen.MAP_ZOOM+Screen.xOffset);
 		}
 		
 		public int getMapY() {
-			return (int)(y/Screen.MAP_SCALE/Screen.MAP_ZOOM+Screen.yOffset);
+			return (int)(((double)y)/Screen.MAP_SCALE/Screen.MAP_ZOOM+Screen.yOffset);
 		}
 		
 		public void refresh(){
